Win32: add divider/font/spacer descriptors, tighten eligibility gate
Descriptor layer:
- Added .divider, .font, .spacer to Win32DescriptorKind
- Added Win32FontDescriptor with Font prop
- winUpdateIntent: .divider -> .none, .font -> .fontStyle, .spacer -> .none
- winApplyFontStyleHook returns .updated(.fontStyle) via winUpdatedHookResult
  (consistent with other descriptive-only intents, not collapsed to keep)

Describable views:
- FontModifiedView: .font kind with font prop + child
- Divider: .divider, no props/children
- Spacer: .spacer, no props/children

Eligibility:
- Added opaque composite rejection (.composite with no children -> false)
- Host gate unchanged: still textContent + colorFill only

Platform fixes:
- TextColorMutationProbe: CGFloat -> padding (CGFloat unavailable on Windows)
- InputEqualityTests: autoreleasepool -> do (unavailable on Windows)

5 new Win32 tests: font/divider/spacer descriptors, font rejects
narrow path, opaque composite rejects narrow path.
307 tests pass.

+++ Examples/Showcase/TextColorMutationProbe/main.swift
@@ -69,7 +69,7 @@ struct TextColorMutationProbeView: View {
                 .foregroundColor(Color(red: 0.45, green: 0.45, blue: 0.45))
 
             Text("Padding = \(padding)")
-                .padding(CGFloat(padding))
+                .padding(padding)
                 .border(Color(red: 0.6, green: 0.6, blue: 0.6))
 
             Divider()
